// BBCJyothi-Prj/src/app/core/models/user-info.ts
export class UserInfo {
    loginName: string;
    wirelessEmail: string;
    user2pageClass: string;
    userAccessIndicator: string;
    sLoginName: string;
    agentId: string;
    status: string;
    equipId: string;
    csLic: string;
    csdeLic: string;
    cqLic: string;
    passwdChg: string;
    lastLogin: string;
    clfoLic: string;
    csLicType: string;
    cqLicType: string;
    csftsdeLic: string;
    webLogin: string;
    sWebLogin: string;
    submitterInd: string;
    userAccess2privclass: string;
    userDefault2wipbin: string;
    supvrDefault2monitor: string;
    user2rcConfig: string;
    user2srvr: string;
    sfaLic: string;
    univLic: string;
    dev: string;
    locale: string;
    nodeId: string;
    offline2privclass: string;
    csftsLic: string;
    cqftsLic: string;
    objid: string;
    userAttuid: string;
    userRole:string;
}